KAN-107 refactor: fetch 로직 axios로 리팩토링
